streamlit-summarizer/pyproject.toml
[project]
name = "oci-content-summary"
version = "0.1.0"
description = "Streamlit app for summarizing content using OCI Generative AI"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.54.0",
    "oci>=2.167.2",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "playwright",
    "pytest-playwright",
    "pytest-timeout",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/config", "src/components", "src/ui", "src/utils"]

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests (requires E2E_REAL_OCI=1)",
]

[tool.pyright]
extraPaths = [".", "src", "tests"]
exclude = ["config", "keys", ".venv", ".sisyphus"]
executionEnvironments = [
  { root = ".", extraPaths = ["tests", "src"] }
]

[tool.basedpyright]
extraPaths = [".", "src", "tests"]
exclude = ["config", "keys", ".venv", ".sisyphus"]
executionEnvironments = [
  { root = ".", extraPaths = ["tests", "src"] }
]
